requirements.txt
snowflake-snowpark-python
streamlit
pypdf
langchain-snowflake
langchain_community
langchain_core
tomli
